etl_job/requirements.txt
pymongo
sqlalchemy
psycopg2-binary
pandas
python-dotenv